adjusted according to suggestion from main code author

+++ degiro_connector/quotecast/tools/ticker_to_df.py
@@ -119,8 +119,8 @@ class TickerToDF:
         df = df.with_columns(
             request_duration_s=pl.col("product_id")
             .map_elements(lambda x: self.__stored_request_duration_map.get(x, None),
-                          return_dtype=pl.Float64) # this prevents map_elements to throw MapWithoutReturnDtypeWarning
-            .cast(pl.Float64)
+                          return_dtype=pl.Int64)
+            .cast(pl.Int64)
         )
 
         return df
@@ -139,7 +139,7 @@ class TickerToDF:
         df = df.with_columns(
             response_datetime=pl.col("product_id").map_elements(
                 lambda x: self.__stored_response_datetime_map.get(x, None),
-                return_dtype=pl.Datetime # this prevents map_elements to throw MapWithoutReturnDtypeWarning
+                return_dtype=pl.Datetime
             )
         )
         df = df.with_columns(
